refactor: Break down App.js into smaller components
Refactors the main App component by breaking it down into smaller, more manageable components for each section of the portfolio. This improves readability, maintainability, and customizability.

- Creates a `src/components` directory for the new components.
- Moves hardcoded data for skills and experience to a separate `src/data.js` file.
- Extracts the CSS from an inline `<style>` tag to a dedicated `src/App.css` file.
- Creates individual components for Header, Hero, About, Skills, Experience, Contact, and Footer sections.
- Updates `App.js` to act as a container component, composing the UI from the new smaller components.
