+++ src/App.js
@@ -1,20 +1,25 @@
 import React, { useState, useEffect } from 'react';
-import { Linkedin, Mail, Phone, MapPin, ChevronDown, Award, Briefcase, GraduationCap, Code, Database, Cloud, BarChart, Brain, GitBranch, Zap, Lightbulb, User, Building } from 'lucide-react';
+import './App.css';
+import Header from './components/Header';
+import Hero from './components/Hero';
+import About from './components/About';
+import Skills from './components/Skills';
+import Experience from './components/Experience';
+import InterestSection from './components/InterestSection';
+import Contact from './components/Contact';
+import Footer from './components/Footer';
 
-// Main App Component
 const App = () => {
     const [loading, setLoading] = useState(true);
     const [activeSection, setActiveSection] = useState('home');
 
-    // Simulate loading time for the page
     useEffect(() => {
         const timer = setTimeout(() => {
             setLoading(false);
-        }, 1500); // Simulate a 1.5 second load time
+        }, 1500);
         return () => clearTimeout(timer);
     }, []);
 
-    // Scroll to section logic
     const scrollToSection = (id) => {
         const element = document.getElementById(id);
         if (element) {
@@ -33,1298 +38,17 @@ const App = () => {
 
     return (
         <div className="app-container">
-            <style>
-                {`
-                @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&display=swap');
-
-                :root {
-                    --purple-600: #8b5cf6;
-                    --indigo-800: #4338ca;
-                    --gray-50: #f9fafb;
-                    --gray-800: #1f2937;
-                    --white: #ffffff;
-                    --gray-900: #111827;
-                    --purple-700: #6d28d9;
-                    --yellow-300: #fde047;
-                    --gray-100: #f3f4f6;
-                    --purple-500: #a78bfa;
-                    --blue-700: #1d4ed8;
-                    --blue-800: #1e40af;
-                    --red-600: #dc2626;
-                    --red-700: #b91c1c;
-                    --green-600: #16a34a;
-                    --green-700: #15803d;
-                    --gray-600: #4b5563;
-                    --gray-700: #374151;
-                    --gray-300: #d1d5db;
-                    --purple-100: #ede9fe;
-                    --purple-800: #6b21a8;
-                    --gray-200: #e5e7eb;
-                    --gray-500: #6b7280;
-                    --cbd5e0: #cbd5e0; /* light text for dark footer */
-                    --2d3748: #2d3748; /* darker blue-gray for footer */
-                }
-
-                body {
-                    font-family: 'Inter', sans-serif;
-                    background-color: var(--gray-50);
-                    color: var(--gray-800);
-                    margin: 0;
-                    overflow-x: hidden;
-                }
-
-                .app-container {
-                    min-height: 100vh;
-                    display: flex;
-                    flex-direction: column;
-                }
-
-                /* Loading Overlay */
-                .loading-overlay-container {
-                    position: fixed;
-                    inset: 0;
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                    background: linear-gradient(135deg, var(--purple-600) 0%, var(--indigo-800) 100%);
-                    z-index: 50;
-                }
-
-                .spinner {
-                    border: 4px solid rgba(255, 255, 255, 0.3);
-                    border-top: 4px solid var(--white);
-                    border-radius: 50%;
-                    width: 50px;
-                    height: 50px;
-                    animation: spin 1s linear infinite;
-                }
-
-                @keyframes spin {
-                    0% { transform: rotate(0deg); }
-                    100% { transform: rotate(360deg); }
-                }
-
-                /* Header */
-                .header {
-                    position: fixed;
-                    top: 0;
-                    left: 0;
-                    width: 100%;
-                    background-color: var(--white);
-                    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-                    z-index: 40;
-                    padding: 1rem 1.5rem; /* py-4 px-6 */
-                    display: flex;
-                    justify-content: space-between;
-                    align-items: center;
-                    border-bottom-left-radius: 0.75rem; /* rounded-b-xl */
-                    border-bottom-right-radius: 0.75rem; /* rounded-b-xl */
-                }
-
-                .header-left {
-                    display: flex;
-                    align-items: center;
-                    gap: 0.5rem; /* space-x-2 */
-                }
-
-                .header-title {
-                    font-size: 1.5rem; /* text-2xl */
-                    font-weight: 700; /* font-bold */
-                    color: var(--gray-900);
-                }
-
-                .header-nav {
-                    display: none; /* hidden */
-                    gap: 2rem; /* space-x-8 */
-                }
-
-                .header-nav-button {
-                    font-size: 1.125rem; /* text-lg */
-                    font-weight: 500; /* font-medium */
-                    color: var(--gray-700);
-                    transition-property: color;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                    background: none;
-                    border: none;
-                    cursor: pointer;
-                    padding-bottom: 0.25rem;
-                }
-
-                .header-nav-button:hover {
-                    color: var(--purple-600);
-                }
-
-                .header-nav-button.active {
-                    color: var(--purple-600);
-                    border-bottom: 2px solid var(--purple-600);
-                }
-
-                .mobile-menu-button {
-                    display: block; /* md:hidden */
-                    color: var(--gray-700);
-                    background: none;
-                    border: none;
-                    cursor: pointer;
-                }
-
-                .mobile-menu-button:hover {
-                    color: var(--purple-600);
-                }
-
-                /* Main Content Area */
-                .main-content {
-                    flex-grow: 1;
-                    padding-top: 5rem; /* pt-20, adjust for fixed header */
-                }
-
-                /* Hero Section */
-                .hero-section {
-                    position: relative;
-                    height: 100vh;
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                    background: linear-gradient(135deg, var(--purple-600) 0%, var(--indigo-800) 100%);
-                    color: var(--white);
-                    overflow: hidden;
-                    text-align: center;
-                }
-
-                .hero-background-dots {
-                    position: absolute;
-                    inset: 0;
-                    z-index: 0;
-                }
-
-                .hero-background-dots div {
-                    position: absolute;
-                    width: 100%;
-                    height: 100%;
-                    background-image: radial-gradient(var(--white) 1px, transparent 1px);
-                    background-size: 20px 20px;
-                    opacity: 0.1;
-                }
-
-                .hero-background-dots .animate-pulse-slow {
-                    animation: pulseSlow 4s infinite;
-                }
-
-                .hero-background-dots .delay-500 {
-                    animation-delay: 0.5s;
-                }
-
-                .hero-content {
-                    z-index: 10;
-                    padding: 1.5rem; /* p-6 */
-                    max-width: 48rem; /* max-w-4xl */
-                    margin-left: auto;
-                    margin-right: auto;
-                }
-
-                .hero-title {
-                    font-size: 3.75rem; /* text-6xl */
-                    font-weight: 800; /* font-extrabold */
-                    line-height: 1.25; /* leading-tight */
-                    margin-bottom: 1rem; /* mb-4 */
-                    animation: fadeInUp 0.6s ease-out forwards;
-                }
-
-                .hero-subtitle {
-                    font-size: 1.875rem; /* text-3xl */
-                    font-weight: 300; /* font-light */
-                    margin-bottom: 2rem; /* mb-8 */
-                    animation: fadeInUp 0.6s ease-out 0.2s forwards; /* delay-200 */
-                }
-
-                .hero-button {
-                    background-color: var(--white);
-                    color: var(--purple-700);
-                    padding: 1rem 2rem; /* px-8 py-4 */
-                    border-radius: 9999px; /* rounded-full */
-                    font-size: 1.25rem; /* text-xl */
-                    font-weight: 600; /* font-semibold */
-                    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); /* shadow-xl */
-                    transition-property: all;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                    cursor: pointer;
-                    border: none;
-                    animation: fadeInUp 0.6s ease-out 0.4s forwards; /* delay-400 */
-                }
-
-                .hero-button:hover {
-                    background-color: var(--gray-100);
-                    transform: scale(1.05);
-                }
-
-                /* Section Styling (General) */
-                .section-padding {
-                    padding-top: 5rem; /* py-20 */
-                    padding-bottom: 5rem; /* py-20 */
-                    padding-left: 1.5rem; /* px-6 */
-                    padding-right: 1.5rem; /* px-6 */
-                }
-
-                .section-container {
-                    max-width: 72rem; /* max-w-6xl */
-                    margin-left: auto;
-                    margin-right: auto;
-                }
-
-                .section-title {
-                    font-size: 2.25rem; /* text-4xl */
-                    font-weight: 700; /* font-bold */
-                    text-align: center;
-                    margin-bottom: 3rem; /* mb-12 */
-                    color: var(--gray-900);
-                    position: relative;
-                }
-
-                .section-title span.relative-text {
-                    position: relative;
-                    z-index: 10;
-                }
-
-                .section-title span.underline-effect {
-                    position: absolute;
-                    left: 50%;
-                    transform: translateX(-50%);
-                    bottom: 0;
-                    height: 0.5rem; /* h-2 */
-                    width: 6rem; /* w-24 */
-                    background-color: var(--purple-500);
-                    z-index: 0;
-                    border-radius: 9999px; /* rounded-full */
-                }
-
-                /* About Section Specific */
-                .about-section {
-                    background-color: var(--white);
-                    box-shadow: inset 0 2px 4px 0 rgba(0, 0, 0, 0.06); /* shadow-inner */
-                }
-
-                .about-grid {
-                    display: grid;
-                    grid-template-columns: 1fr;
-                    gap: 3rem; /* gap-12 */
-                    align-items: center;
-                }
-
-                .about-text {
-                    font-size: 1.125rem; /* text-lg */
-                    line-height: 1.75; /* leading-relaxed */
-                    animation: fadeInLeft 0.6s ease-out forwards;
-                }
-
-                .about-text p {
-                    margin-bottom: 1rem; /* mb-4 */
-                }
-
-                .education-card {
-                    background-color: var(--gray-100);
-                    padding: 2rem; /* p-8 */
-                    border-radius: 0.75rem; /* rounded-xl */
-                    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* shadow-md */
-                    animation: fadeInRight 0.6s ease-out forwards;
-                }
-
-                .education-card-title {
-                    font-size: 1.5rem; /* text-2xl */
-                    font-weight: 600; /* font-semibold */
-                    margin-bottom: 1.5rem; /* mb-6 */
-                    color: var(--purple-700);
-                    display: flex;
-                    align-items: center;
-                }
-
-                .education-list {
-                    list-style: none;
-                    padding: 0;
-                    margin: 0;
-                    display: flex;
-                    flex-direction: column;
-                    gap: 1rem; /* space-y-4 */
-                }
-
-                .education-list-item {
-                    display: flex;
-                    align-items: flex-start;
-                }
-
-                .education-list-item p {
-                    margin: 0;
-                }
-
-                .education-list-item .title {
-                    font-weight: 500; /* font-medium */
-                    color: var(--gray-900);
-                }
-
-                .education-list-item .detail {
-                    color: var(--gray-600);
-                }
-
-                /* Skills Section Specific */
-                .skills-section {
-                    background-color: var(--gray-100);
-                }
-
-                .skills-grid {
-                    display: grid;
-                    grid-template-columns: repeat(1, 1fr);
-                    gap: 2rem; /* gap-8 */
-                }
-
-                .skill-category-card {
-                    background-color: var(--white);
-                    padding: 1.5rem; /* p-6 */
-                    border-radius: 0.75rem; /* rounded-xl */
-                    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* shadow-md */
-                    transition-property: transform;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                    animation: fadeInUp 0.5s ease-out forwards; /* animate-fade-in */
-                }
-
-                .skill-category-card:hover {
-                    transform: scale(1.05);
-                }
-
-                .skill-category-header {
-                    display: flex;
-                    align-items: center;
-                    margin-bottom: 1rem; /* mb-4 */
-                    color: var(--purple-700);
-                }
-
-                .skill-category-title {
-                    font-size: 1.25rem; /* text-xl */
-                    font-weight: 600; /* font-semibold */
-                }
-
-                .skill-tags {
-                    display: flex;
-                    flex-wrap: wrap;
-                    gap: 0.5rem; /* gap-2 */
-                }
-
-                .skill-tag {
-                    background-color: var(--purple-100);
-                    color: var(--purple-800);
-                    font-size: 0.875rem; /* text-sm */
-                    font-weight: 500; /* font-medium */
-                    padding: 0.25rem 0.75rem; /* px-3 py-1 */
-                    border-radius: 9999px; /* rounded-full */
-                    box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05); /* shadow-sm */
-                    transition-property: background-color;
-                    transition-duration: 200ms;
-                    transition-timing-function: ease;
-                    cursor: pointer;
-                }
-
-                .skill-tag:hover {
-                    background-color: var(--purple-200);
-                }
-
-                /* Experience Section Specific */
-                .experience-section {
-                    background-color: var(--white);
-                    box-shadow: inset 0 2px 4px 0 rgba(0, 0, 0, 0.06); /* shadow-inner */
-                }
-
-                .job-experience-container {
-                    display: flex;
-                    flex-direction: column;
-                    gap: 2.5rem; /* space-y-10 */
-                }
-
-                .job-card {
-                    background-color: var(--white);
-                    padding: 2rem; /* p-8 */
-                    border-radius: 0.75rem; /* rounded-xl */
-                    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); /* shadow-lg */
-                    border: 1px solid var(--gray-100);
-                    animation: fadeInUp 0.6s ease-out forwards;
-                }
-
-                .job-card-header {
-                    display: flex;
-                    align-items: center;
-                    justify-content: space-between;
-                    cursor: pointer;
-                }
-
-                .job-card-header-left {
-                    display: flex;
-                    align-items: center;
-                }
-
-                .job-title {
-                    font-size: 1.5rem; /* text-2xl */
-                    font-weight: 700; /* font-bold */
-                    color: var(--gray-900);
-                }
-
-                .job-company {
-                    font-size: 1.125rem; /* text-lg */
-                    color: var(--gray-600);
-                }
-
-                .job-duration {
-                    font-size: 1rem; /* text-md */
-                    color: var(--gray-500);
-                }
-
-                .job-chevron {
-                    color: var(--gray-500);
-                    transition-property: transform;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                }
-
-                .job-chevron.rotated {
-                    transform: rotate(180deg);
-                }
-
-                .job-projects-content {
-                    margin-top: 1.5rem; /* mt-6 */
-                    padding-top: 1.5rem; /* pt-6 */
-                    border-top: 1px solid var(--gray-200);
-                }
-
-                .job-projects-title {
-                    font-size: 1.25rem; /* text-xl */
-                    font-weight: 600; /* font-semibold */
-                    margin-bottom: 1rem; /* mb-4 */
-                    color: var(--purple-700);
-                }
-
-                .job-projects-list {
-                    list-style: none;
-                    padding: 0;
-                    margin: 0;
-                    display: flex;
-                    flex-direction: column;
-                    gap: 1rem; /* space-y-4 */
-                    color: var(--gray-700);
-                }
-
-                .job-project-item {
-                    position: relative;
-                    padding-left: 1.5rem; /* pl-6 */
-                }
-
-                .job-project-item::before {
-                    content: '';
-                    position: absolute;
-                    left: 0;
-                    top: 0.375rem; /* top-1.5 */
-                    width: 0.5rem; /* w-2 */
-                    height: 0.5rem; /* h-2 */
-                    background-color: var(--purple-500);
-                    border-radius: 50%;
-                }
-
-                .job-project-item strong {
-                    font-weight: 600; /* font-semibold */
-                }
-
-                .job-project-item .impact-text {
-                    display: block;
-                    font-size: 0.875rem; /* text-sm */
-                    color: var(--gray-500);
-                    font-style: italic;
-                    margin-top: 0.25rem; /* mt-1 */
-                }
-
-                /* Company Interest Form Section */
-                .interest-section {
-                    background-color: var(--gray-100);
-                }
-
-                .interest-form-container {
-                    max-width: 32rem; /* max-w-xl */
-                    margin-left: auto;
-                    margin-right: auto;
-                    text-align: center;
-                    background-color: var(--white);
-                    padding: 2rem; /* p-8 */
-                    border-radius: 0.75rem; /* rounded-xl */
-                    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); /* shadow-lg */
-                }
-
-                .interest-form-title {
-                    font-size: 2.25rem; /* text-4xl */
-                    font-weight: 700; /* font-bold */
-                    margin-bottom: 1.5rem; /* mb-6 */
-                    color: var(--gray-900);
-                    position: relative;
-                }
-
-                .interest-form-title span.relative-text {
-                    position: relative;
-                    z-index: 10;
-                }
-
-                .interest-form-title span.underline-effect {
-                    position: absolute;
-                    left: 50%;
-                    transform: translateX(-50%);
-                    bottom: 0;
-                    height: 0.5rem; /* h-2 */
-                    width: 6rem; /* w-24 */
-                    background-color: var(--purple-500);
-                    z-index: 0;
-                    border-radius: 9999px; /* rounded-full */
-                }
-
-                .interest-form-description {
-                    font-size: 1.125rem; /* text-lg */
-                    color: var(--gray-700);
-                    margin-bottom: 2rem; /* mb-8 */
-                }
-
-                .form-layout { /* Added for overall form layout */
-                    display: flex;
-                    flex-direction: column;
-                    gap: 1.5rem; /* space-y-6 equivalent */
-                    text-align: left; /* Ensure content is left-aligned within the form */
-                }
-
-                .form-group {
-                    margin-bottom: 0; /* Remove default margin-bottom if gap is used on parent */
-                }
-
-                .form-group label {
-                    display: block;
-                    color: var(--gray-700);
-                    font-size: 0.875rem; /* text-sm */
-                    font-weight: 500; /* font-medium */
-                    margin-bottom: 0.5rem; /* mb-2 */
-                    text-align: left; /* Ensure label is left-aligned */
-                    padding-left: 1rem; /* Align label with input padding */
-                }
-
-                .form-group input,
-                .form-group textarea {
-                    width: 100%;
-                    padding: 0.75rem 1rem; /* px-4 py-2 */
-                    border: 1px solid var(--gray-300);
-                    border-radius: 0.5rem; /* rounded-lg */
-                    outline: none;
-                    box-sizing: border-box; /* Include padding and border in the element's total width and height */
-                }
-
-                .form-group input:focus,
-                .form-group textarea:focus {
-                    border-color: var(--purple-500);
-                    box-shadow: 0 0 0 2px rgba(139, 92, 246, 0.5); /* focus:ring-2 focus:ring-purple-500 */
-                }
-
-                .form-submit-button {
-                    width: 100%;
-                    background-color: var(--purple-600);
-                    color: var(--white);
-                    padding: 0.75rem 1.5rem; /* px-6 py-3 */
-                    border-radius: 9999px; /* rounded-full */
-                    font-size: 1.125rem; /* text-lg */
-                    font-weight: 600; /* font-semibold */
-                    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* shadow-md */
-                    transition-property: background-color;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                    border: none;
-                    cursor: pointer;
-                }
-
-                .form-submit-button:hover {
-                    background-color: var(--purple-700);
-                }
-
-                /* Contact Section Specific */
-                .contact-section {
-                    background-color: var(--gray-50);
-                }
-
-                .contact-container {
-                    max-width: 64rem; /* max-w-4xl */
-                    margin-left: auto;
-                    margin-right: auto;
-                    text-align: center;
-                }
-
-                .contact-links {
-                    display: flex;
-                    flex-direction: column;
-                    justify-content: center;
-                    align-items: center;
-                    /* Removed gap: 2rem; and margin-bottom: 3rem; to allow for dynamic spacing */
-                    margin-bottom: 2rem; /* Adjusted for fewer items */
-                }
-
-                .contact-link-item {
-                    display: flex;
-                    align-items: center;
-                    gap: 0.75rem; /* space-x-3 */
-                    background-color: var(--white);
-                    padding: 1.5rem; /* p-6 */
-                    border-radius: 0.75rem; /* rounded-xl */
-                    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* shadow-md */
-                    transition-property: all;
-                    transition-duration: 300ms;
-                    transition-timing-function: ease;
-                    text-decoration: none; /* remove underline from links */
-                    margin-bottom: 1rem; /* Add margin between items */
-                }
-
-                .contact-link-item:hover {
-                    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); /* hover:shadow-lg */
-                    transform: scale(1.05);
-                }
-
-                .contact-link-item span {
-                    font-size: 1.125rem; /* text-lg */
-                    font-weight: 500; /* font-medium */
-                    color: var(--gray-700);
-                    transition-property: color;
-                    transition-duration: 300ms;
-                }
-
-                .contact-link-item:hover span {
-                    color: var(--gray-900);
-                }
-
-                .contact-address {
-                    display: none; /* Hidden as requested */
-                }
-
-                /* Footer */
-                .footer {
-                    background-color: var(--gray-900);
-                    color: var(--gray-300);
-                    padding-top: 2rem; /* py-8 */
-                    padding-bottom: 2rem; /* py-8 */
-                    text-align: center;
-                    margin-top: auto; /* flex-grow in main-content */
-                    border-top-left-radius: 0.75rem; /* rounded-t-xl */
-                    border-top-right-radius: 0.75rem; /* rounded-t-xl */
-                }
-
-                .footer-container {
-                    max-width: 72rem; /* max-w-6xl */
-                    margin-left: auto;
-                    margin-right: auto;
-                }
-
-                .footer p {
-                    margin: 0;
-                }
-
-                .footer .small-text {
-                    display: none; /* Hidden as requested */
-                }
-
-                /* Media Queries for Responsiveness */
-                @media (min-width: 768px) { /* md breakpoint */
-                    .header-nav {
-                        display: flex; /* md:flex */
-                    }
-                    .mobile-menu-button {
-                        display: none; /* md:hidden */
-                    }
-                    .about-grid {
-                        grid-template-columns: repeat(2, 1fr); /* md:grid-cols-2 */
-                    }
-                    .skills-grid {
-                        grid-template-columns: repeat(2, 1fr); /* md:grid-cols-2 */
-                    }
-                    .contact-links {
-                        flex-direction: row; /* md:flex-row */
-                        gap: 2rem; /* md:space-x-8 */
-                        margin-bottom: 3rem; /* md:space-y-0 */
-                    }
-                    .contact-link-item {
-                        margin-bottom: 0; /* Remove margin between items for row layout */
-                    }
-                }
-
-                @media (min-width: 1024px) { /* lg breakpoint */
-                    .skills-grid {
-                        grid-template-columns: repeat(3, 1fr); /* lg:grid-cols-3 */
-                    }
-                }
-
-                /* Keyframe Animations (defined in JS for consistency, but here for CSS-only context) */
-                @keyframes fadeInUp {
-                  from { opacity: 0; transform: translateY(20px); }
-                  to { opacity: 1; transform: translateY(0); }
-                }
-                @keyframes fadeInLeft {
-                  from { opacity: 0; transform: translateX(-20px); }
-                  to { opacity: 1; transform: translateX(0); }
-                }
-                @keyframes fadeInRight {
-                  from { opacity: 0; transform: translateX(20px); }
-                  to { opacity: 1; transform: translateX(0); }
-                }
-                @keyframes pulseSlow {
-                  0%, 100% { opacity: 0.1; }
-                  50% { opacity: 0.2; }
-                }
-
-                .animate-fade-in-up { animation: fadeInUp 0.6s ease-out forwards; }
-                .animate-fade-in-up.delay-200 { animation-delay: 0.2s; }
-                .animate-fade-in-up.delay-400 { animation-delay: 0.4s; }
-                .animate-fade-in-left { animation: fadeInLeft 0.6s ease-out forwards; }
-                .animate-fade-in-right { animation: fadeInRight 0.6s ease-out forwards; }
-                .animate-fade-in { animation: fadeInUp 0.5s ease-out forwards; }
-                .animate-pulse-slow { animation: pulseSlow 4s infinite; }
-                `}
-            </style>
-
-            {/* Header */}
-            <header className="header">
-                <div className="header-left">
-                    <User style={{ color: 'var(--purple-600)' }} size={28} />
-                    <h1 className="header-title">Sudharsana Rajasekaran</h1>
-                </div>
-                <nav className="header-nav">
-                    {['home', 'about', 'skills', 'experience', 'contact', 'interest'].map((section) => (
-                        <button
-                            key={section}
-                            onClick={() => scrollToSection(section)}
-                            className={`header-nav-button ${activeSection === section ? 'active' : ''}`}
-                        >
-                            {section.charAt(0).toUpperCase() + section.slice(1)}
-                        </button>
-                    ))}
-                </nav>
-                {/* Mobile Menu Button - Add a proper mobile menu if needed */}
-                <div className="mobile-menu-button-wrapper">
-                    <button className="mobile-menu-button">
-                        <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M4 6h16M4 12h16m-7 6h7"></path>
-                        </svg>
-                    </button>
-                </div>
-            </header>
-
+            <Header scrollToSection={scrollToSection} activeSection={activeSection} />
             <main className="main-content">
-                {/* Hero Section */}
-                <section id="home" className="hero-section">
-                    <div className="hero-background-dots">
-                        <div className="animate-pulse-slow" style={{ top: 0, left: 0 }}></div>
-                        <div className="animate-pulse-slow delay-500" style={{ bottom: 0, right: 0 }}></div>
-                    </div>
-                    <div className="hero-content">
-                        <h2 className="hero-title animate-fade-in-up">
-                            Hello, I'm <span style={{ color: 'var(--yellow-300)' }}>Sudharsana Rajasekaran</span>
-                        </h2>
-                        <p className="hero-subtitle animate-fade-in-up delay-200">
-                            A <span style={{ fontWeight: '600' }}>Senior Data Engineer</span> passionate about building scalable data solutions and driving business insights.
-                        </p>
-                        <button
-                            onClick={() => scrollToSection('experience')}
-                            className="hero-button animate-fade-in-up delay-400"
-                        >
-                            View My Projects
-                        </button>
-                    </div>
-                </section>
-
-                {/* About Section */}
-                <section id="about" className="section-padding about-section">
-                    <div className="section-container">
-                        <h3 className="section-title">
-                            <span className="relative-text">About Me</span>
-                            <span className="underline-effect"></span>
-                        </h3>
-                        <div className="about-grid">
-                            <div className="about-text animate-fade-in-left">
-                                <p>
-                                    I am a highly accomplished Senior Data Engineer with a proven track record of architecting and implementing robust data pipelines that drive significant business impact. My expertise lies in transforming complex data into actionable insights, leading to tangible improvements in revenue, operational efficiency, and fraud prevention.
-                                </p>
-                                <p>
-                                    With a Master's in Business Analytics (Data Science) and a background in Electronics & Communication Engineering, I bring a unique blend of technical prowess and business acumen. I thrive in collaborative environments, ensuring seamless synergy between engineering and product development to achieve broader organizational objectives.
-                                </p>
-                                <p>
-                                    My career highlights include spearheading data initiatives at Walmart Global Tech and building critical BI solutions for Google's global data center operations. I am passionate about leveraging cutting-edge technologies and data science techniques to solve real-world problems.
-                                </p>
-                            </div>
-                            <div className="education-card animate-fade-in-right">
-                                <h4 className="education-card-title">
-                                    <GraduationCap style={{ marginRight: '0.75rem' }} size={24} /> Education
-                                </h4>
-                                <ul className="education-list">
-                                    <li className="education-list-item">
-                                        <Award style={{ color: 'var(--yellow-500)', marginRight: '0.75rem', marginTop: '0.25rem' }} size={20} />
-                                        <div>
-                                            <p className="title">M.S in Business Analytics (Data Science)</p>
-                                            <p className="detail">The University of Texas at Dallas</p>
-                                        </div>
-                                    </li>
-                                    <li className="education-list-item">
-                                        <Award style={{ color: 'var(--yellow-500)', marginRight: '0.75rem', marginTop: '0.25rem' }} size={20} />
-                                        <div>
-                                            <p className="title">B.E., Electronics & Communication Engineering</p>
-                                            <p className="detail">Anna University</p>
-                                        </div>
-                                    </li>
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
-                </section>
-
-                {/* Skills Section */}
-                <section id="skills" className="section-padding skills-section">
-                    <div className="section-container">
-                        <h3 className="section-title">
-                            <span className="relative-text">Technical Skills</span>
-                            <span className="underline-effect"></span>
-                        </h3>
-                        <div className="skills-grid">
-                            {[
-                                { category: 'Programming', icon: Code, skills: ['Python', 'Scala', 'Java'] },
-                                { category: 'Big Data', icon: Zap, skills: ['Apache Hadoop', 'Hive', 'Apache Kafka', 'Apache Spark', 'Apache Airflow', 'Apache NiFi'] },
-                                { category: 'Databases', icon: Database, skills: ['Presto', 'Cosmos DB', 'Microsoft SQL Server', 'MySQL', 'Oracle', 'Teradata', 'MongoDB'] },
-                                { category: 'Cloud Technologies', icon: Cloud, skills: ['Google Cloud Platform (GCP)', 'Google Cloud Storage (GCS)', 'Google BigQuery', 'Google DataProc'] },
-                                { category: 'Data Science Skills', icon: Brain, skills: ['Regression Techniques', 'Neural Networks', 'ANOVA', 'Hypothesis Testing', 'Time Series Analysis', 'Data Mining', 'Data Visualization', 'Predictive Analytics', 'Deep Learning', 'Computer Vision'] },
-                                { category: 'BI/ETL/Reporting Tools', icon: BarChart, skills: ['Tableau', 'Power BI', 'Looker', 'SAP Business Objects', 'IBM Cognos'] },
-                                { category: 'CI/CD Automation', icon: GitBranch, skills: ['Looper'] },
-                            ].map((skillCategory, index) => (
-                                <div key={index} className="skill-category-card animate-fade-in" style={{ animationDelay: `${index * 0.1}s` }}>
-                                    <div className="skill-category-header">
-                                        {React.createElement(skillCategory.icon, { size: 28, style: { marginRight: '0.75rem' } })}
-                                        <h4 className="skill-category-title">{skillCategory.category}</h4>
-                                    </div>
-                                    <div className="skill-tags">
-                                        {skillCategory.skills.map((skill, skillIndex) => (
-                                            <span key={skillIndex} className="skill-tag">
-                                                {skill}
-                                            </span>
-                                        ))}
-                                    </div>
-                                </div>
-                            ))}
-                        </div>
-                    </div>
-                </section>
-
-                {/* Experience Section */}
-                <section id="experience" className="section-padding experience-section">
-                    <div className="section-container">
-                        <h3 className="section-title">
-                            <span className="relative-text">Professional Experience & Projects</span>
-                            <span className="underline-effect"></span>
-                        </h3>
-                        <div className="job-experience-container">
-                            {[
-                                {
-                                    title: 'Senior Data Engineer',
-                                    company: 'Walmart Global Tech, Sunnyvale, California',
-                                    duration: 'March 2022 - Present',
-                                    icon: Briefcase,
-                                    projects: [
-                                        {
-                                            name: 'Automated Data Pipelines & Customer Data Marts',
-                                            description: 'Spearheaded the identification and integration of valuable data sources, architecting and implementing automated data pipelines utilizing Spark, Scala, and Airflow, enriching customer-specific data marts to serve diverse business needs.',
-                                            impact: 'Improved data accessibility and insights for various business functions.'
-                                        },
-                                        {
-                                            name: 'Affiliate Data Pipeline for Walmart Business',
-                                            description: 'Independently designed, built, and deployed an affiliate data pipeline for Walmart Business.',
-                                            impact: 'Resulted in 4.9M new user visits and a $5.4M increase in GMV within the first year of production.'
-                                        },
-                                        {
-                                            name: 'Real-time Fraud Prevention Pipeline',
-                                            description: 'Engineered a real-time data pipeline incorporating advanced validation mechanisms to detect and prevent fraudulent sign-up activity.',
-                                            impact: 'Resulted in $30,000 in average daily savings.'
-                                        },
-                                        {
-                                            name: 'Walmart Business & Salesforce Integration',
-                                            description: 'Orchestrated the consolidation of multifaceted Walmart Business (WMB) customer data, synchronizing customer information with Salesforce to provide a unified, comprehensive view.',
-                                            impact: 'Facilitated informed decision-making for the WMB sales team and streamlined operational efficiency.'
-                                        },
-                                        {
-                                            name: 'Customer Events & Profiles Analysis',
-                                            description: 'Conducted in-depth analysis and reporting on customer events and profiles, offering strategic recommendations for site enhancements and contributing to root cause analysis efforts.',
-                                            impact: 'Supported business and product initiatives, leading to data-driven improvements.'
-                                        },
-                                        {
-                                            name: 'NiFi/Spark Data Pipelines for Revenue Growth',
-                                            description: 'Delivered $1M+ in new monthly revenue by implementing NiFi/Spark data pipelines to integrate Walmart Business and Salesforce systems.',
-                                            impact: 'Enabled automated customer communication for new and expanding business opportunities, capturing high-value, GMV-driving contracts.'
-                                        }
-                                    ]
-                                },
-                                {
-                                    title: 'Business Intelligence Engineer',
-                                    company: 'Google (Contractor), Sunnyvale, California',
-                                    duration: 'August 2020 - March 2022',
-                                    icon: Lightbulb,
-                                    projects: [
-                                        {
-                                            name: 'BI Solution for Google Data Center Supply Chain',
-                                            description: 'Built and managed a BI solution to visualize purchase orders and vendor invoices for worldwide Google Data Center supply chain operations.',
-                                            impact: 'Provided critical visibility into supply chain financials.'
-                                        },
-                                        {
-                                            name: 'Python Script Optimization (HDFS to Google Sheets)',
-                                            description: 'Optimized a Python script to run 10x faster, used across Google to load data from HDFS to Google Sheets.',
-                                            impact: 'Significantly improved data loading efficiency.'
-                                        },
-                                        {
-                                            name: 'Python ETL Framework for Data Center Planning',
-                                            description: 'Implemented a Python ETL framework to automate the process of extracting and cleansing data from 700+ data collection sheets into a BI Dashboard.',
-                                            impact: 'Helped visualize and plan new data center construction operations.'
-                                        },
-                                        {
-                                            name: 'Data Analysis for Infrastructure Optimization',
-                                            description: 'Performed various Data Analysis using SQL and Python.',
-                                            impact: 'Optimized and improved the efficiency of Google Data Center infrastructure and operations.'
-                                        },
-                                        {
-                                            name: 'Tableau Dashboards for Self-Service Reports',
-                                            description: 'Built Tableau Dashboards to create self-service reports for various stakeholders.',
-                                            impact: 'Empowered stakeholders to make strategic business decisions independently.'
-                                        },
-                                        {
-                                            name: 'Data Object & Report Migration',
-                                            description: 'Facilitated the migration of 13 data objects, retrofitted 48 reports & 40+ data tables, and enabling integration of 4 boundary systems.',
-                                            impact: 'Ensured seamless transition and continued data availability.'
-                                        }
-                                    ]
-                                },
-                                {
-                                    title: 'Data Analyst Intern',
-                                    company: 'Ericsson, Plano, Texas',
-                                    duration: 'September 2019 - May 2020',
-                                    icon: Briefcase,
-                                    projects: [
-                                        {
-                                            name: 'Automated Analytical Reports',
-                                            description: 'Extracted data from various sites and automated analytical reports to manage 11 teams.',
-                                            impact: 'Saved 12-15 hours per week in manual reporting efforts.'
-                                        },
-                                        {
-                                            name: 'Analytics Platform for Business Recommendations',
-                                            description: 'Designed and implemented analytics platform to provide recommendations to strategic business specific to Ericsson North America.',
-                                            impact: 'Enabled data-driven strategic decision-making.'
-                                        },
-                                        {
-                                            name: 'Network Performance Data Transformation',
-                                            description: 'Transformed network performance data utilizing open-source datasets and provided performance metric insights for customers.',
-                                            impact: 'Optimized network product and maintenance based on performance data.'
-                                        },
-                                        {
-                                            name: 'Web Scraping & Data Integration',
-                                            description: 'Extracted data from third-party websites by Web scraping, web crawling and integrated it with existing tower data in the data lake.',
-                                            impact: 'Avoided data discrepancies and thereby saving revenue.'
-                                        },
-                                        {
-                                            name: 'TensorFlow Models for 5G Implementations',
-                                            description: 'Deployed TensorFlow models to detect and identify cars, objects, and poles for 5G implementations.',
-                                            impact: 'Enhanced capabilities for 5G infrastructure development.'
-                                        },
-                                        {
-                                            name: 'Image Auto-Labeling with Feature Matching',
-                                            description: 'Enforced auto labeling of images by using Feature matching technique.',
-                                            impact: 'Saved manual effort hours by 50%.'
-                                        }
-                                    ]
-                                },
-                                {
-                                    title: 'Business Intelligence Consultant',
-                                    company: 'Cognizant Technology Solutions, Chennai',
-                                    duration: 'June 2014 - June 2018',
-                                    icon: Briefcase,
-                                    projects: [
-                                        {
-                                            name: 'Use Case Scenarios & Database Queries',
-                                            description: 'Designed use case scenarios, complex database queries and test plans for generating business insights across various clients such as automobile, pharmaceutical, retail and entertainment.',
-                                            impact: 'Provided foundational insights for diverse client needs.'
-                                        },
-                                        {
-                                            name: 'SQL ETL Framework for Chat Helpdesk',
-                                            description: 'Built an SQL ETL framework to parse unstructured data into a data table that can be better visualized on a dashboard.',
-                                            impact: 'Improved the chat helpdesk system overall ticket resolution count by approximately 30%.'
-                                        },
-                                        {
-                                            name: 'Client & Development Team Coordination',
-                                            description: 'Coordinated with clients and development teams to translate business requirements into strategic solutions.',
-                                            impact: 'Ensured alignment between business needs and technical implementation.'
-                                        },
-                                        {
-                                            name: 'Business Automation & Contract Initiation',
-                                            description: 'Initiated contracts with clients and identified opportunities to develop and automate existing business.',
-                                            impact: 'Drove new business opportunities and operational efficiencies.'
-                                        },
-                                        {
-                                            name: 'Team Leadership & BI Reporting',
-                                            description: 'Led a team of 10 developers for a project of $5M and delivered month-over-month BI reports and dashboards.',
-                                            impact: 'Successfully managed large-scale BI projects and reporting cycles.'
-                                        },
-                                        {
-                                            name: 'Mobile Dashboards & SAP BO Migration',
-                                            description: 'Conceptualized and created three mobile dashboards and facilitated the migration of SAP BO reports to new versions.',
-                                            impact: 'Enhanced mobile accessibility of insights and modernized reporting infrastructure.'
-                                        }
-                                    ]
-                                }
-                            ].map((job, jobIndex) => (
-                                <JobExperience key={jobIndex} job={job} />
-                            ))}
-                        </div>
-                    </div>
-                </section>
-
-                {/* Company Interest Form Section */}
-                <section id="interest" className="section-padding interest-section">
-                    <div className="interest-form-container">
-                        <h3 className="interest-form-title">
-                            <span className="relative-text">Interested in Collaborating?</span>
-                            <span className="underline-effect"></span>
-                        </h3>
-                        <p className="interest-form-description">
-                            If your company is interested in discussing potential opportunities or collaborations, please fill out the form below.
-                        </p>
-                        <CompanyInterestForm />
-                    </div>
-                </section>
-
-                {/* Contact Section */}
-                <section id="contact" className="section-padding contact-section">
-                    <div className="contact-container">
-                        <h3 className="section-title">
-                            <span className="relative-text">Get in Touch</span>
-                            <span className="underline-effect"></span>
-                        </h3>
-                        <div className="contact-links">
-                            <a href="https://www.linkedin.com/in/sudharsanarajasekaran/" target="_blank" rel="noopener noreferrer" className="contact-link-item">
-                                <Linkedin style={{ color: 'var(--blue-700)' }} size={32} />
-                                <span>LinkedIn Profile</span>
-                            </a>
-                            {/* Email removed as requested */}
-                            {/* Phone number removed as requested */}
-                        </div>
-                        {/* Address removed as requested */}
-                    </div>
-                </section>
+                <Hero scrollToSection={scrollToSection} />
+                <About />
+                <Skills />
+                <Experience />
+                <InterestSection />
+                <Contact />
             </main>
-
-            {/* Footer */}
-            <footer className="footer">
-                <div className="footer-container">
-                    <p>&copy; {new Date().getFullYear()} Sudharsana Rajasekaran. All rights reserved.</p>
-                    {/* Removed "Built with React & Plain CSS" as requested */}
-                </div>
-            </footer>
+            <Footer />
         </div>
-    );
-};
-
-// Component for individual job experience
-const JobExperience = ({ job }) => {
-    const [isOpen, setIsOpen] = useState(false);
-
-    return (
-        <div className="job-card animate-fade-in-up">
-            <div className="job-card-header" onClick={() => setIsOpen(!isOpen)}>
-                <div className="job-card-header-left">
-                    {React.createElement(job.icon, { size: 32, style: { color: 'var(--purple-600)', marginRight: '1rem' } })}
-                    <div>
-                        <h4 className="job-title">{job.title}</h4>
-                        <p className="job-company">{job.company}</p>
-                        <p className="job-duration">{job.duration}</p>
-                    </div>
-                </div>
-                <ChevronDown className={`job-chevron ${isOpen ? 'rotated' : ''}`} size={24} />
-            </div>
-            {isOpen && (
-                <div className="job-projects-content">
-                    <h5 className="job-projects-title">Key Projects & Contributions:</h5>
-                    <ul className="job-projects-list">
-                        {job.projects.map((project, projectIndex) => (
-                            <li key={projectIndex} className="job-project-item">
-                                <strong>{project.name}:</strong> {project.description}
-                                {project.impact && <span className="impact-text">Impact: {project.impact}</span>}
-                            </li>
-                        ))}
-                    </ul>
-                </div>
-            )}
-        </div>
-    );
-};
-
-// New Component for Company Interest Form
-const CompanyInterestForm = () => {
-    const [formData, setFormData] = useState({
-        companyName: '',
-        contactPerson: '',
-        email: '',
-        message: ''
-    });
-
-    const handleChange = (e) => {
-        const { name, value } = e.target;
-        setFormData(prevState => ({
-            ...prevState,
-            [name]: value
-        }));
-    };
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        // In a real application, you would send this data to a backend service.
-        // For example, a Google Cloud Function, App Engine endpoint, or a third-party form service.
-        console.log('Form submitted:', formData);
-
-        // Display a simple message box instead of alert()
-        const messageBox = document.createElement('div');
-        messageBox.className = 'fixed-message-box-overlay';
-        messageBox.innerHTML = `
-            <div class="fixed-message-box-content">
-                <h4 class="message-box-title">Thank You!</h4>
-                <p class="message-box-text">Your interest has been noted. Sudharsana will get back to you soon.</p>
-                <button id="closeMessageBox" class="message-box-button">Close</button>
-            </div>
-        `;
-        document.body.appendChild(messageBox);
-
-        document.getElementById('closeMessageBox').addEventListener('click', () => {
-            document.body.removeChild(messageBox);
-        });
-
-        // Clear the form
-        setFormData({
-            companyName: '',
-            contactPerson: '',
-            email: '',
-            message: ''
-        });
-    };
-
-    return (
-        <form onSubmit={handleSubmit} className="form-layout">
-            <div className="form-group">
-                <label htmlFor="companyName">Company Name</label>
-                <input
-                    type="text"
-                    id="companyName"
-                    name="companyName"
-                    value={formData.companyName}
-                    onChange={handleChange}
-                    placeholder="Your Company Name"
-                    required
-                />
-            </div>
-            <div className="form-group">
-                <label htmlFor="contactPerson">Contact Person</label>
-                <input
-                    type="text"
-                    id="contactPerson"
-                    name="contactPerson"
-                    value={formData.contactPerson}
-                    onChange={handleChange}
-                    placeholder="Your Name"
-                    required
-                />
-            </div>
-            <div className="form-group">
-                <label htmlFor="email">Company Email</label>
-                <input
-                    type="email"
-                    id="email"
-                    name="email"
-                    value={formData.email}
-                    onChange={handleChange}
-                    placeholder="contact@company.com"
-                    required
-                />
-            </div>
-            <div className="form-group">
-                <label htmlFor="message">Message / Area of Interest</label>
-                <textarea
-                    id="message"
-                    name="message"
-                    value={formData.message}
-                    onChange={handleChange}
-                    rows="5"
-                    placeholder="Tell us about your interest..."
-                    required
-                ></textarea>
-            </div>
-            <button
-                type="submit"
-                className="form-submit-button"
-            >
-                <Building style={{ marginRight: '0.5rem' }} size={20} /> Submit Interest
-            </button>
-            <style>
-                {`
-                .form-layout {
-                    display: flex;
-                    flex-direction: column;
-                    gap: 1.5rem; /* space-y-6 equivalent */
-                    text-align: left; /* Ensure content is left-aligned within the form */
-                }
-
-                .fixed-message-box-overlay {
-                    position: fixed;
-                    inset: 0;
-                    background-color: rgba(0, 0, 0, 0.5);
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                    z-index: 50;
-                }
-
-                .fixed-message-box-content {
-                    background-color: var(--white);
-                    padding: 2rem;
-                    border-radius: 0.5rem;
-                    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
-                    text-align: center;
-                    max-width: 24rem;
-                    margin-left: auto;
-                    margin-right: auto;
-                }
-
-                .message-box-title {
-                    font-size: 1.5rem;
-                    font-weight: 700;
-                    color: var(--gray-900);
-                    margin-bottom: 1rem;
-                }
-
-                .message-box-text {
-                    color: var(--gray-700);
-                    margin-bottom: 1.5rem;
-                }
-
-                .message-box-button {
-                    background-color: var(--purple-600);
-                    color: var(--white);
-                    padding: 0.5rem 1.5rem;
-                    border-radius: 9999px;
-                    transition-property: background-color;
-                    transition-duration: 300ms;
-                    cursor: pointer;
-                    border: none;
-                }
-
-                .message-box-button:hover {
-                    background-color: var(--purple-700);
-                }
-                `}
-            </style>
-        </form>
     );
 };
