+++ src/App.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import App from './App';
 
-test('renders learn react link', () => {
+test('renders the main heading after loading', async () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+  const headingElement = await screen.findByRole('heading', { name: /Hello, I'm Sudharsana Rajasekaran/i, level: 2 }, { timeout: 3000 });
+  expect(headingElement).toBeInTheDocument();
 });
